Ultra (#257)
* style: remove rich text editor max width

* feat: aspect ratio component

* feat: capsture component with delete

* feat: admin crops module

* fix: other crops link to

* feat: about us admin form

* feat: home admin form

* feat: privacy admin form

* feat terms admin form

* kwakwakwakwak

* feat: crops and farm problems paths in admin

* feat: crops admin get data view

* feat: create crops hook

* feat: update crops hook

* feat: farm problems inc learning resource

* style: fix design bug landing

* style: notification max height scroll

* feat: tags module admin

* feat: crops module admin wii

* feat: radio component shad

* style: remove updatedat column

* feat: display and add, update companion

* stye: datatable to table crops list

* style: parse html elements in display

* fix: update map correspond to new api

* to fix files

* style: change farm management icon in admin

* feat: navigate to profile from admin header

* style: yieldable checkbox style v1.1

* int: client details to footer

* style: log out confirmation dialog

* style: signup final setup tu

* style: move forgot password below

* feat: add see more

* style: basta may binago dito pagod na ko

* int: client details

* int: client details

* style: redesign CTA v4

* style: carousel to see all

* style: edited to hardcoded

* style: display no recent post

* style: display no saved yet

* feat: get client details hook

* fix get tags hook

* style: back button in farm appication

* stye: adjust spacing between 2 graphs

* style: redesign resource oveview

* style: login back button

* feat: loader, backbutton, related

* style: last updated at

* style: add redirect to contact

* dom purify to parse blogs list

* dompurify to farm blogs view

* ayoko na

* style: redesign change logo

* style: add neggative mt

+++ src/components/ui/custom/header/user-header-menu.tsx
@@ -16,6 +16,17 @@ import ActivityIndicator from "@icons/ActivityIndicator";
 import { RiAdminLine } from "react-icons/ri";
 import { useNavigate } from "react-router-dom";
 import { Link } from "react-router-dom";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger
+} from "@components/ui/alert-dialog";
 
 const UserHeaderMenu = () => {
   const navigate = useNavigate();
@@ -83,16 +94,33 @@ const UserHeaderMenu = () => {
               <span className="font-poppins-medium">Settings</span>
             </DropdownMenuItem>
           </Link>
-
-          <DropdownMenuItem
-            className="cursor-pointer h-12 gap-2"
-            onClick={() => {
-              deleteAuthData();
-            }}
-          >
-            <MdLogout className="text-foreground/80 text-lg " />
-            <span className="font-poppins-medium">Logout</span>
-          </DropdownMenuItem>
+          <DropdownMenuSeparator />
+          <AlertDialog>
+            <AlertDialogTrigger asChild>
+              <div className="flex gap-2 p-3">
+                <MdLogout className="text-foreground/80 text-lg " />
+                <span className="font-poppins-medium">Logout</span>
+              </div>
+            </AlertDialogTrigger>
+            <AlertDialogContent>
+              <AlertDialogHeader>
+                <AlertDialogTitle>Log out your account?</AlertDialogTitle>
+                <AlertDialogDescription>
+                  This will log out your account from this device.
+                </AlertDialogDescription>
+              </AlertDialogHeader>
+              <AlertDialogFooter>
+                <AlertDialogCancel>Cancel</AlertDialogCancel>
+                <AlertDialogAction
+                  onClick={() => {
+                    deleteAuthData();
+                  }}
+                >
+                  Continue
+                </AlertDialogAction>
+              </AlertDialogFooter>
+            </AlertDialogContent>
+          </AlertDialog>
         </DropdownMenuContent>
       </DropdownMenu>
